// backend/src/cases/dto/query-cases.dto.ts
export class QueryCasesDto {
  search?: string;
  status?: string;
  isClosed?: string;
  priority?: string;
  department?: string;
  personInCharge?: string;
  moduleLevel?: string;
  dateFrom?: string;
  dateTo?: string;
  page?: string;
  pageSize?: string;
  sortField?: string;
  sortDir?: string;
}